from master

+++ app/Ninja/Repositories/BaseRepository.php
@@ -1,25 +1,46 @@
 <?php namespace App\Ninja\Repositories;
 
+/**
+ * Class BaseRepository
+ */
 class BaseRepository
 {
+    /**
+     * @return null
+     */
     public function getClassName() 
     {
         return null;
     }
 
+    /**
+     * @return mixed
+     */
     private function getInstance()
     {
         $className = $this->getClassName();
         return new $className();
     }
 
+    /**
+     * @param $entity
+     * @param $type
+     * @return string
+     */
     private function getEventClass($entity, $type)
     {
         return 'App\Events\\' . ucfirst($entity->getEntityType()) . 'Was' . $type;
     }
 
+    /**
+     * @param $entity
+     */
     public function archive($entity)
     {
+        if ($entity->trashed()) {
+            return;
+        }
+        
         $entity->delete();
 
         $className = $this->getEventClass($entity, 'Archived');
@@ -29,8 +50,15 @@ class BaseRepository
         }
     }
 
+    /**
+     * @param $entity
+     */
     public function restore($entity)
     {
+        if ( ! $entity->trashed()) {
+            return;
+        }
+
         $fromDeleted = false;
         $entity->restore();
 
@@ -47,8 +75,15 @@ class BaseRepository
         }
     }
 
+    /**
+     * @param $entity
+     */
     public function delete($entity)
     {
+        if ($entity->is_deleted) {
+            return;
+        }
+        
         $entity->is_deleted = true;
         $entity->save();
 
@@ -61,11 +96,19 @@ class BaseRepository
         }
     }
 
+    /**
+     * @param $ids
+     * @return mixed
+     */
     public function findByPublicIds($ids)
     {
         return $this->getInstance()->scope($ids)->get();
     }
 
+    /**
+     * @param $ids
+     * @return mixed
+     */
     public function findByPublicIdsWithTrashed($ids)
     {
         return $this->getInstance()->scope($ids)->withTrashed()->get();
